% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRMSE.R
\name{calcRMSE}
\alias{calcRMSE}
\title{Title calcRMSE}
\usage{
calcRMSE(mech, size, sdat.c, pythonPath, wd)
}
\arguments{
\item{mech}{Missing data mechanism of interest (e.g. MCAR,MAR,MNAR)}

\item{size}{Size of dataset you want to compare imputation methods on}

\item{sdat.c}{Dataset to sample from for plasmode simulations}

\item{pythonPath}{specify where on computer python is (desired version <3.8)}

\item{wd}{set working directory (should contain AEDatsets folder and python script)}
}
\value{
Writes csv with calculated RMSE for DAE and MICE
#' @import MASS
}
\description{
Title calcRMSE
}
\examples{
Data <- generateData(100)
 workingdir <-  "C:/Users/kgetz1/Documents/Year2/ComputingProject/GitHub/CompareMI"
path <-  "C:/Users/kgetz1/AppData/Local/Programs/Python/Python38/python.exe"
calcRMSE("MCAR",size=100,sdat.c=Data,pythonPath=path,wd=workingdir)

}
