% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generateData}
\alias{generateData}
\title{Generate Data}
\usage{
generateData(size)
}
\arguments{
\item{size}{Size of dataset you want to draw from when doing plasmode simulations}
}
\value{
Survival dataset of desired size with approximately 20\% censored data with 10 variables
}
\description{
Generate Data
}
\examples{
generateData(100)
generateData(1000)


}
