% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResults.R
\name{plotResults}
\alias{plotResults}
\title{Title plotResults}
\usage{
plotResults(mech, lowerlim, upperlim, size, wd)
}
\arguments{
\item{mech}{Missing data mechanism of interest (e.g. MCAR,MAR,MNAR)}

\item{lowerlim}{Lower limit for plotting}

\item{upperlim}{Upper limit for plotting}

\item{size}{Size of dataset you are comparing imputation methods on}

\item{wd}{Specify working directory (should contain results from RestructureResults)}
}
\value{
plots results comparing imputation methods and saves pdfs
}
\description{
Title plotResults
}
\examples{
 workingdir <-  "C:/Users/kgetz1/Documents/Year2/ComputingProject/GitHub/CompareMI"
plotResults("MCAR",-4,4,size=100,wd=workingdir)

}
