% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestructureResults.R
\name{RestructureResults}
\alias{RestructureResults}
\title{RestructureResults}
\usage{
RestructureResults(mech, size, wd)
}
\arguments{
\item{mech}{Missing data mechanism of interest (e.g. MCAR,MAR,MNAR)}

\item{size}{Size of dataset you want to compare imputation methods on}

\item{wd}{Specify working directory (should contain AEDatsets folder and python script)}
}
\value{
Writes csv files with reorganized structure to prepare for plotting
}
\description{
RestructureResults
}
\examples{
 workingdir <-  "C:/Users/kgetz1/Documents/Year2/ComputingProject/GitHub/CompareMI"
RestructureResults(mech="MCAR",size=100,wd=workingdir)

}
