% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSims.R
\name{runSims}
\alias{runSims}
\title{runSims}
\usage{
runSims(
  mech,
  size,
  iterations,
  Data,
  parallelOption,
  coreNum = 1,
  pythonPath,
  wd
)
}
\arguments{
\item{mech}{Missing data mechanism of interest (e.g. MCAR,MAR,MNAR)}

\item{size}{Size of dataset you want to compare imputation methods on}

\item{iterations}{Number of iterations for simulation}

\item{Data}{Dataset to sample from for plasmode simulations}

\item{parallelOption}{Set equal to T if you want to parallelize}

\item{coreNum}{Specify number of core if parallelizing}

\item{pythonPath}{specify where on computer python is (desired version <3.8)}

\item{wd}{set working directory (should contain AEDatsets folder and python script)}
}
\value{
Writes files with information about bias and percentiles for each imputation method
}
\description{
runSims
}
\examples{
sdat.c <- generateData(100)
path <-  "C:/Users/kgetz1/AppData/Local/Programs/Python/Python38/python.exe"
workingdir <-  "C:/Users/kgetz1/Documents/Year2/ComputingProject/GitHub/CompareMI"
runSims("MCAR",size=100,iterations=2,Data=sdat.c,parallelOption=FALSE,pythonPath=path,wd=workingdir)
}
